Removed forced while loop
Seems like telegram's API will ban you from sending images if you send too many. In that case, it returns an exception. Need to look into it more, but for now, I've modified it slightly so that if telegram does not allow you to send an image, the image will not go through, just the message.